+++ src/main/java/com/ketheroth/uncrafter/common/inventory/container/InputHandler.java
@@ -1,16 +1,11 @@
 package com.ketheroth.uncrafter.common.inventory.container;
 
-import com.ketheroth.uncrafter.common.config.Configuration;
-import net.minecraft.world.item.EnchantedBookItem;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.crafting.Recipe;
-import net.minecraft.world.item.enchantment.EnchantmentHelper;
-import net.minecraft.world.item.enchantment.EnchantmentInstance;
 import net.minecraftforge.items.ItemStackHandler;
 
 import javax.annotation.Nonnull;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 import java.util.Random;
 
@@ -30,12 +25,6 @@ public class InputHandler extends ItemStackHandler {
 		for (int i = 0; i < 9; i++) {
 			container.getOutputHandler().setStackInSlot(i, ItemStack.EMPTY);
 		}
-		if (container.isAdvanced() && container.getEnchantmentHandler() != null) {
-			for (int i = 0; i < 6; i++) {
-				container.getEnchantmentHandler().setStackInSlot(i, ItemStack.EMPTY);
-			}
-			container.getEnchantmentHandler().resetExtracting();
-		}
 		if (!stack.is(container.getCache().getA())) {
 			Recipe<?> recipe = UncrafterContainer.searchRecipe(stack, container.getRecipeManager());
 			if (recipe != null) {
@@ -54,19 +43,6 @@ public class InputHandler extends ItemStackHandler {
 			container.getOutputHandler().setStackInSlot(index, ingredient.copy());
 			index++;
 		}
-		//add enchantments
-		if (container.isAdvanced() && container.getEnchantmentHandler() != null && stack.isEnchanted()) {
-			ArrayList<ItemStack> books = new ArrayList<>();
-			EnchantmentHelper.getEnchantments(stack).forEach((enchantment, level) -> {
-				if (!enchantment.isCurse()) {
-					books.add(EnchantedBookItem.createForEnchantment(new EnchantmentInstance(enchantment, Configuration.MINIMUM_LEVEL_FOR_ENCHANTMENTS.get() ? 1 : level)));
-				}
-			});
-			Collections.shuffle(books);
-			for (int i = 0; i < 6 && i < books.size(); i++) {
-				container.getEnchantmentHandler().setStackInSlot(i, books.get(i));
-			}
-		}
 		return super.insertItem(slot, stack, simulate);
 	}
 
@@ -80,12 +56,6 @@ public class InputHandler extends ItemStackHandler {
 		if (!simulate) {
 			for (int i = 0; i < 9; i++) {
 				container.getOutputHandler().setStackInSlot(i, ItemStack.EMPTY);
-			}
-			if (container.isAdvanced() && container.getEnchantmentHandler() != null) {
-				for (int i = 0; i < 6; i++) {
-					container.getEnchantmentHandler().setStackInSlot(i, ItemStack.EMPTY);
-				}
-				container.getEnchantmentHandler().resetExtracting();
 			}
 			if (this.getStackInSlot(slot).getCount() > amount) {
 				ItemStack stack = this.getStackInSlot(0);
@@ -105,14 +75,6 @@ public class InputHandler extends ItemStackHandler {
 				for (ItemStack ingredient : container.getCache().getB()) {
 					container.getOutputHandler().setStackInSlot(index, ingredient.copy());
 					index++;
-				}
-				if (container.isAdvanced() && container.getEnchantmentHandler() != null && stack.isEnchanted()) {
-					ArrayList<ItemStack> books = new ArrayList<>();
-					EnchantmentHelper.getEnchantments(stack).forEach((enchantment, level) -> books.add(EnchantedBookItem.createForEnchantment(new EnchantmentInstance(enchantment, Configuration.MINIMUM_LEVEL_FOR_ENCHANTMENTS.get() ? 1 : level))));
-					Collections.shuffle(books);
-					for (int i = 0; i < 6 && i < books.size(); i++) {
-						container.getEnchantmentHandler().setStackInSlot(i, books.get(i));
-					}
 				}
 			}
 		}